Fix: deprecated functions and refactor
